+++ src/bin/app.js
@@ -29,6 +29,7 @@ import compare from '../middleware/compare.js'
 import Hook from '../utils/hook.js'
 import Biz from '../utils/biz.js'
 import { doCommand } from '../utils/process.js'
+import { loadPlugin } from '../utils/plugin.js'
 
 /*-----------------
   about router
@@ -83,35 +84,6 @@ if(fs.existsSync(configPath)){
   config = _.assign(config, require(configPath));
 }
 
-const loadPlugin = function(fpm){
-  let modulesDir = path.join(CWD, 'node_modules')
-  let files = fs.readdirSync(modulesDir)
-  files = _.filter(files, (f)=>{
-    return _.startsWith(f, 'fpm-plugin-')
-  })
-  //load package.json
-  let plugins = {}
-  _.map(files, (f)=>{
-    let packageInfo = require(path.join(modulesDir, f, 'package.json'))
-    plugins[packageInfo.name] = { name: packageInfo.name, version: packageInfo.version, info: packageInfo, npm:`https://www.npmjs.com/package/${packageInfo.name}`, registry: `http://registry.npmjs.org/${packageInfo.name}` }
-  })
-  _.map(plugins, (p) => {
-    let m = require(path.join(modulesDir, p.name))
-    if(_.has(m, 'default')){
-      m = m.default
-    }
-    if(_.isFunction(m.getDependencies)){
-      let deps = m.getDependencies() || []
-      _.map(deps, (d) => {
-        if(!_.has(plugins, d.name)){
-          throw new Exception(E.System.PLUGIN_LOAD_ERROR(p.name, d.name))
-        }
-      })
-    }
-    p.ref = m.bind(fpm)
-  })
-  fpm._plugins = _.assign(fpm._plugins, plugins)
-}
 
 class Fpm {
   constructor(){
@@ -136,10 +108,8 @@ class Fpm {
     this._prject_info = projectInfo
     this._env = config.dev
     this._version = packageInfo.version
-    this._plugins = {}
-
     //add plugins
-    loadPlugin(this)
+    this._plugins = loadPlugin(this) || {}
     this.runAction('INIT', this)
     this.errorHandler = (err, ctx) => {
     	this.logger.error('server error', err, ctx)